+++ lib/src/style/EventFilters.cpp
@@ -3,6 +3,7 @@
 
 #include "EventFilters.hpp"
 
+#include <oclero/qlementine/style/QlementineStyle.hpp>
 #include <oclero/qlementine/utils/StateUtils.hpp>
 #include <oclero/qlementine/utils/ImageUtils.hpp>
 #include <oclero/qlementine/utils/PrimitiveUtils.hpp>
@@ -23,6 +24,7 @@
 #include <QComboBox>
 #include <QListView>
 #include <QLayout>
+#include <QPlainTextEdit>
 
 namespace oclero::qlementine {
 LineEditButtonEventFilter::LineEditButtonEventFilter(
@@ -245,17 +247,10 @@ bool TabBarEventFilter::eventFilter(QObject* watchedObject, QEvent* evt) {
         return true;
       }
     }
-
-    // Hack!
-    // QTabBar.cpp, line 1478
-    // We need the QTabBar to set d->layoutDirty to true. The only we I found was to
-    // call QTabBar::setIconSize() because it doesn't check if the icon size is different
-    // before forcing a whole refresh.
-    _tabBar->setIconSize(_tabBar->iconSize());
   } else if (type == QEvent::Wheel) {
     const auto* wheelEvent = static_cast<QWheelEvent*>(evt);
 
-    // Block non-horizontal scorll.
+    // Block non-horizontal scroll.
     const bool wheelVertical = qAbs(wheelEvent->angleDelta().y()) > qAbs(wheelEvent->angleDelta().x());
     if (wheelVertical) {
       evt->ignore();
@@ -513,8 +508,10 @@ bool WidgetWithFocusFrameEventFilter::eventFilter(QObject* watchedObject, QEvent
       if (_focusFrame == nullptr) {
         // Create the focus frame as late as possible to give
         // more chances to any parent scroll area to already exist.
-        _focusFrame = new QFocusFrame(_widget);
-        _focusFrame->setWidget(_widget);
+        QTimer::singleShot(0, this, [this]() {
+          _focusFrame = new QFocusFrame(_widget);
+          _focusFrame->setWidget(_widget);
+        });
       }
     }
   }
@@ -528,7 +525,7 @@ class LineEditMenuIconsBehavior : public QObject {
 
   enum class IconListMode {
     None,
-    LineEdit,
+    LineEdit, // Use of QLineEdit and QPlainTextEdit.
     ReadOnlyLineEdit,
     SpinBox,
   };
@@ -583,11 +580,15 @@ class LineEditMenuIconsBehavior : public QObject {
   }
 
   static IconListMode getMode(const QMenu* menu) {
-    if (const auto* menu_parent = menu->parent()) {
-      if (qobject_cast<const QAbstractSpinBox*>(menu_parent->parent())) {
+    if (const auto* menuParent = menu->parent()) {
+      if (qobject_cast<const QAbstractSpinBox*>(menuParent->parent())) {
         return IconListMode::SpinBox;
-      } else if (const auto* line_edit = qobject_cast<const QLineEdit*>(menu_parent)) {
+      } else if (const auto* line_edit = qobject_cast<const QLineEdit*>(menuParent)) {
         return line_edit->isReadOnly() ? IconListMode::ReadOnlyLineEdit : IconListMode::LineEdit;
+      } else if (const auto* menuParentParent = menuParent ? menuParent->parent() : nullptr) {
+        if (const auto* plainTextEdit = qobject_cast<const QPlainTextEdit*>(menuParentParent)) {
+          return plainTextEdit->isReadOnly() ? IconListMode::ReadOnlyLineEdit : IconListMode::LineEdit;
+        }
       }
     }
     return IconListMode::None;
@@ -625,9 +626,11 @@ public:
 LineEditMenuEventFilter::LineEditMenuEventFilter(QWidget* parent)
   : QObject(parent) {
   assert(parent);
+  // Might be a menu's submenu.
   if (auto* menu = qobject_cast<QMenu*>(parent)) {
     new LineEditMenuIconsBehavior(menu);
   } else {
+    // The QLineEdit.
     parent->installEventFilter(this);
   }
 }
@@ -635,11 +638,36 @@ LineEditMenuEventFilter::LineEditMenuEventFilter(QWidget* parent)
 bool LineEditMenuEventFilter::eventFilter(QObject*, QEvent* evt) {
   const auto type = evt->type();
   if (type == QEvent::ChildPolished) {
+    constexpr auto propertyName = "qlementine_tweak_menu_icons";
     auto* child = static_cast<QChildEvent*>(evt)->child();
-    if (auto* lineedit = qobject_cast<QLineEdit*>(child)) {
-      lineedit->installEventFilter(this);
-    } else if (auto* menu = qobject_cast<QMenu*>(child)) {
-      new LineEditMenuIconsBehavior(menu);
+
+    const auto tweaked = child->property(propertyName).toBool();
+    if (!tweaked) {
+      child->setProperty(propertyName, true);
+
+      // QLineEdit child of QSpinBox.
+      if (auto* lineEdit = qobject_cast<QLineEdit*>(child)) {
+        lineEdit->installEventFilter(this);
+
+      }
+      // Qmenu that needs tweaking.
+      else if (auto* menu = qobject_cast<QMenu*>(child)) {
+        new LineEditMenuIconsBehavior(menu);
+
+        // Forward auto icon color mode from parent to the menu.
+        if (const auto* menuParent = menu->parentWidget()) {
+          if (const auto* style = qobject_cast<oclero::qlementine::QlementineStyle*>(menuParent->style())) {
+            const auto autoIconColor = style->autoIconColor(menuParent);
+            QlementineStyle::setAutoIconColor(menu, autoIconColor);
+          }
+        }
+      }
+      // Case of a QPlainTextEdit (inherits QAbstractScrollArea).
+      else if (child->objectName() == "qt_scrollarea_viewport") {
+        if (auto* childWidget = qobject_cast<QWidget*>(child)) {
+          childWidget->installEventFilter(this);
+        }
+      }
     }
   }