+++ lib/src/style/QlementineStyle.cpp
@@ -55,6 +55,7 @@
 #include <QTextEdit>
 #include <QSpinBox>
 #include <QFontComboBox>
+#include <QTreeView>
 
 #include <cmath>
 #include <mutex>
@@ -480,7 +481,7 @@ void QlementineStyle::drawPrimitive(PrimitiveElement pe, const QStyleOption* opt
         const auto isFlat = optButton->features.testFlag(QStyleOptionButton::Flat);
         const auto mouse = isFlat ? getToolButtonMouseState(opt->state) : getMouseState(opt->state);
         const auto role = getColorRole(opt->state, isDefault);
-        const auto& bgColor = buttonBackgroundColor(mouse, role, w);
+        const auto& bgColor = isFlat ? toolButtonBackgroundColor(mouse, role) : buttonBackgroundColor(mouse, role, w);
         const auto& currentBgColor =
           _impl->animations.animateBackgroundColor(w, bgColor, _impl->theme.animationDuration);
         const auto radiuses = optRoundedButton ? optRoundedButton->radiuses : RadiusesF{ _impl->theme.borderRadius };
@@ -578,14 +579,20 @@ void QlementineStyle::drawPrimitive(PrimitiveElement pe, const QStyleOption* opt
       return;
     case PE_PanelLineEdit:
       if (const auto* optPanelLineEdit = qstyleoption_cast<const QStyleOptionFrame*>(opt)) {
-        const auto* parentWidget = w->parentWidget();
+        const auto* parentWidget = w ? w->parentWidget() : nullptr;
         const auto* parentParentWidget = parentWidget ? parentWidget->parentWidget() : nullptr;
         const auto isTabCellEditor =
           parentParentWidget && qobject_cast<const QAbstractItemView*>(parentParentWidget->parentWidget());
 
+        const auto isComboBoxLineEdit = qobject_cast<const QComboBox*>(w->parentWidget());
+        const auto qPlainTextEdit = qobject_cast<const QPlainTextEdit*>(w);
+        const auto isPlainQPlainTextEdit = qPlainTextEdit && qPlainTextEdit->frameShadow() == QFrame::Shadow::Plain;
+        const auto isPlainLineEdit = !isComboBoxLineEdit && !qPlainTextEdit && optPanelLineEdit->lineWidth == 0;
+        const auto isPlain = isPlainQPlainTextEdit || isPlainLineEdit;
+
         const auto radiusF = static_cast<double>(_impl->theme.borderRadius);
         auto radiuses = RadiusesF{ radiusF };
-        if (isTabCellEditor || w->metaObject()->className() == QStringLiteral("QExpandingLineEdit")) {
+        if (isPlain || isTabCellEditor || (w && w->metaObject()->className() == QStringLiteral("QExpandingLineEdit"))) {
           // The QExpandingLineEdit class is used by QStyleItemDelegate when the cell context type is text.
           radiuses.topRight = 0.;
           radiuses.bottomRight = 0.;
@@ -613,7 +620,9 @@ void QlementineStyle::drawPrimitive(PrimitiveElement pe, const QStyleOption* opt
 
         // Background.
         drawRoundedRect(p, rect, bgColor, radiuses);
-        drawRoundedRectBorder(p, rect, currentBorderColor, borderW, radiuses);
+        if (!isPlainLineEdit) {
+          drawRoundedRectBorder(p, rect, currentBorderColor, borderW, radiuses);
+        }
       }
       return;
     case PE_IndicatorArrowDown:
@@ -1683,8 +1692,25 @@ void QlementineStyle::drawControl(ControlElement ce, const QStyleOption* opt, QP
       return;
     case CE_HeaderSection:
       if (const auto* optHeader = qstyleoption_cast<const QStyleOptionHeader*>(opt)) {
-        const auto* tableView = (w ? qobject_cast<const QTableView*>(w->parentWidget()) : nullptr);
+        const auto* wParent = w ? w->parentWidget() : nullptr;
+        const auto* tableView = qobject_cast<const QTableView*>(wParent);
+        const auto* treeView = qobject_cast<const QTreeView*>(wParent);
+        const auto* horizontalHeader = tableView  ? tableView->horizontalHeader()
+                                       : treeView ? treeView->header()
+                                                  : nullptr;
+        const auto* verticalHeader = tableView ? tableView->verticalHeader() : nullptr;
+        const auto isVertical = optHeader->orientation == Qt::Vertical;
+        const auto isHorizontal = optHeader->orientation == Qt::Horizontal;
         const auto& rect = opt->rect;
+
+        // Sometimes, we don't want external borders, for aesthetics purposes.
+        // Example: a QTreeView beside a QSplitter. We want to avoid the splitter's separator and the
+        // header's borders being side to side, because it'll look like a larger ugly border.
+        const auto shadow = tableView  ? tableView->frameShadow()
+                            : treeView ? treeView->frameShadow()
+                                       : QFrame::Shadow::Sunken;
+        const auto drawTableExternalBorders = shadow != QFrame::Plain;
+
         // Background.
         const auto mouse = getMouseState(opt->state);
         const auto checked = getCheckState(opt->state);
@@ -1698,37 +1724,54 @@ void QlementineStyle::drawControl(ControlElement ce, const QStyleOption* opt, QP
         p->setBrush(Qt::NoBrush);
         p->setPen(QPen(lineColor, lineW));
 
+        // Line at the top.
+        const auto drawTopBorder =
+          drawTableExternalBorders
+          || (isVertical && optHeader->position == QStyleOptionHeader::SectionPosition::Beginning)
+          || (isHorizontal && optHeader->position != QStyleOptionHeader::SectionPosition::Beginning);
+        if (drawTopBorder) {
+          const auto horizontalHeaderHidden = horizontalHeader ? horizontalHeader->isHidden() : true;
+          if (optHeader->orientation == Qt::Horizontal
+              || (horizontalHeaderHidden && optHeader->position == QStyleOptionHeader::Beginning)) {
+            const auto p1 = QPointF(rect.x(), rect.y() + lineW * .5);
+            const auto p2 = QPointF(rect.x() + rect.width(), rect.y() + lineW * .5);
+            p->drawLine(p1, p2);
+          }
+        }
+
         // Line on the right.
-        /*if (optHeader->position != QStyleOptionHeader::SectionPosition::OnlyOneSection)*/ {
+        const auto drawRightBorder =
+          drawTableExternalBorders || isVertical
+          || (isHorizontal && optHeader->position == QStyleOptionHeader::SectionPosition::End);
+        if (drawRightBorder) {
           const auto p1 = QPointF(rect.x() + rect.width() - lineW * .5, rect.y());
           const auto p2 = QPointF(rect.x() + rect.width() - lineW * .5, rect.y() + rect.height());
           p->drawLine(p1, p2);
         }
 
-        // Line below.
-        {
+        // Line at the bottom.
+        const auto drawBottomBorder = drawTableExternalBorders
+                                      || (isVertical && optHeader->position != QStyleOptionHeader::SectionPosition::End)
+                                      || isHorizontal;
+        if (drawBottomBorder) {
           const auto p1 = QPointF(rect.x(), rect.y() + rect.height() - lineW * .5);
           const auto p2 = QPointF(rect.x() + rect.width(), rect.y() + rect.height() - lineW * .5);
           p->drawLine(p1, p2);
         }
 
-        // Line at the top.
-        const auto horizontalHeaderHidden = tableView ? tableView->horizontalHeader()->isHidden() : true;
-        if (optHeader->orientation == Qt::Horizontal
-            || (horizontalHeaderHidden && optHeader->position == QStyleOptionHeader::Beginning)) {
-          const auto p1 = QPointF(rect.x(), rect.y() + lineW * .5);
-          const auto p2 = QPointF(rect.x() + rect.width(), rect.y() + lineW * .5);
-          p->drawLine(p1, p2);
-        }
-
         // Line at the left.
-        const auto verticalHeaderHidden = tableView ? tableView->verticalHeader()->isHidden() : true;
-        if (optHeader->orientation == Qt::Vertical
-            || (optHeader->orientation == Qt::Horizontal && optHeader->position == QStyleOptionHeader::OnlyOneSection)
-            || (verticalHeaderHidden && optHeader->position == QStyleOptionHeader::Beginning)) {
-          const auto p1 = QPointF(rect.x() + lineW * .5, rect.y());
-          const auto p2 = QPointF(rect.x() + lineW * .5, rect.y() + rect.height());
-          p->drawLine(p1, p2);
+        const auto drawLeftBorder =
+          drawTableExternalBorders
+          || (isHorizontal && optHeader->position == QStyleOptionHeader::SectionPosition::Beginning);
+        if (drawLeftBorder) {
+          const auto verticalHeaderHidden = verticalHeader ? verticalHeader->isHidden() : true;
+          if (optHeader->orientation == Qt::Vertical
+              || (optHeader->orientation == Qt::Horizontal && optHeader->position == QStyleOptionHeader::OnlyOneSection)
+              || (verticalHeaderHidden && optHeader->position == QStyleOptionHeader::Beginning)) {
+            const auto p1 = QPointF(rect.x() + lineW * .5, rect.y());
+            const auto p2 = QPointF(rect.x() + lineW * .5, rect.y() + rect.height());
+            p->drawLine(p1, p2);
+          }
         }
       }
       return;
@@ -1825,11 +1868,20 @@ void QlementineStyle::drawControl(ControlElement ce, const QStyleOption* opt, QP
     case CE_SizeGrip:
       break;
     case CE_Splitter: {
+      constexpr auto maxSplitterThickness = 2;
+      constexpr auto minSplitterThickness = 1;
+      const auto& rect = opt->rect;
       const auto mouse = getMouseState(opt->state);
       const auto& lineColor = splitterColor(mouse);
-      // const auto currentLineColor = _impl->animations.animateBackgroundColor(w, lineColor, _impl->theme.animationDuration);
-      const auto line_rect = opt->rect.adjusted(-1, 0, 1, 0);
-      p->fillRect(line_rect, lineColor);
+      const auto isHorizontal = opt->state.testFlag(QStyle::State_Horizontal);
+      const auto lineThickness =
+        std::clamp(isHorizontal ? rect.width() : rect.height(), minSplitterThickness, maxSplitterThickness);
+      const auto lineW = isHorizontal ? lineThickness : rect.width();
+      const auto lineH = isHorizontal ? rect.height() : lineThickness;
+      const auto lineX = isHorizontal ? rect.x() + (rect.width() - lineThickness) / 2 : rect.x();
+      const auto lineY = isHorizontal ? rect.y() : rect.y() + (rect.height() - lineThickness) / 2;
+      const auto lineRect = QRect(lineX, lineY, lineW, lineH);
+      p->fillRect(lineRect, lineColor);
     }
       return;
     // case CE_RubberBand:
@@ -1954,10 +2006,14 @@ void QlementineStyle::drawControl(ControlElement ce, const QStyleOption* opt, QP
           // Dial: placed around the handle.
           optFocus.rect = subElementRect(SE_SliderFocusRect, &optDial, dial);
           optFocus.radiuses = optFocus.rect.height() / 2.;
-        } else if (qobject_cast<const QLineEdit*>(monitoredWidget)) {
+        } else if (const auto* lineEdit = qobject_cast<const QLineEdit*>(monitoredWidget)) {
           // LineEdit: placed around the whole text field.
           const auto* parentWidget = monitoredWidget ? monitoredWidget->parentWidget() : nullptr;
           const auto* parentParentWidget = parentWidget ? parentWidget->parentWidget() : nullptr;
+
+          // Check if the QLineEdit should have radiuses.
+          const auto isComboBoxLineEdit = qobject_cast<const QComboBox*>(lineEdit->parentWidget()) != nullptr;
+          const auto isPlainLineEdit = !isComboBoxLineEdit && !lineEdit->hasFrame();
 
           // Check if the QLineEdit is a cell editor of a QTableView or equivalent.
           const auto isTabCellEditor =
@@ -1968,13 +2024,13 @@ void QlementineStyle::drawControl(ControlElement ce, const QStyleOption* opt, QP
           const auto* parentSpinbox = qobject_cast<const QAbstractSpinBox*>(parentWidget);
           const auto* parentCombobox = qobject_cast<const QComboBox*>(parentWidget);
 
-          const auto margin = borderW;
+          const auto margin = isPlainLineEdit ? borderW * 2 : borderW;
           optFocus.rect = optFocus.rect.marginsRemoved(QMargins(margin, margin, margin, margin));
           optFocus.radiuses = _impl->theme.borderRadius;
 
           // Check if the QLineEdit is inside a QSpinBox and +/- buttons are visible,
           // or inside an editable QComboBox.
-          if (isTabCellEditor) {
+          if (isPlainLineEdit || isTabCellEditor) {
             optFocus.radiuses.topRight = 0.;
             optFocus.radiuses.topLeft = 0.;
             optFocus.radiuses.bottomRight = 0.;
@@ -2544,7 +2600,7 @@ void QlementineStyle::drawComplexControl(
   switch (cc) {
     case CC_SpinBox:
       if (const auto* spinboxOpt = qstyleoption_cast<const QStyleOptionSpinBox*>(opt)) {
-        const auto* parentWidget = w->parentWidget();
+        const auto* parentWidget = w ? w->parentWidget() : nullptr;
         const auto isTabCellEditor =
           parentWidget && qobject_cast<const QAbstractItemView*>(parentWidget->parentWidget());
 
@@ -2656,7 +2712,7 @@ void QlementineStyle::drawComplexControl(
             }
           }
         } else {
-          const auto* parentWidget = w->parentWidget();
+          const auto* parentWidget = w ? w->parentWidget() : nullptr;
           const auto isTabCellEditor =
             parentWidget && qobject_cast<const QAbstractItemView*>(parentWidget->parentWidget());
 
@@ -3302,22 +3358,14 @@ QRect QlementineStyle::subControlRect(
           } break;
           case SC_ComboBoxEditField: {
             if (comboBoxOpt->editable) {
+              const auto hasIcon = !comboBoxOpt->currentIcon.isNull();
               const auto indicatorSize = _impl->theme.iconSize;
               const auto spacing = _impl->theme.spacing;
-              const auto isBasicComboBox =
-                qobject_cast<const QComboBox*>(w) != nullptr && qobject_cast<const QFontComboBox*>(w) == nullptr;
-              if (isBasicComboBox) {
-                // Strange hack to place the QLineEdit correctly.
-                const auto indicatorButtonW = spacing * 2 + indicatorSize.width();
-                const auto shiftX = static_cast<int>(spacing * 2.5);
-                const auto editFieldW = comboBoxOpt->rect.width() - indicatorButtonW + shiftX;
-                return QRect{ comboBoxOpt->rect.x() - shiftX, comboBoxOpt->rect.y(), editFieldW,
-                  comboBoxOpt->rect.height() };
-              } else {
-                const auto indicatorButtonW = spacing * 2 + indicatorSize.width();
-                const auto editFieldW = comboBoxOpt->rect.width() - indicatorButtonW;
-                return QRect{ comboBoxOpt->rect.x(), comboBoxOpt->rect.y(), editFieldW, comboBoxOpt->rect.height() };
-              }
+              const auto shiftX = hasIcon ? static_cast<int>(spacing * 2.5) : 0;
+              const auto indicatorButtonW = spacing * 2 + indicatorSize.width();
+              const auto editFieldW = comboBoxOpt->rect.width() - indicatorButtonW + shiftX;
+              return QRect{ comboBoxOpt->rect.x() - shiftX, comboBoxOpt->rect.y(), editFieldW,
+                comboBoxOpt->rect.height() };
             } else {
               return QRect{};
             }
@@ -3461,7 +3509,7 @@ QRect QlementineStyle::subControlRect(
         const auto menuIsOnSeparateButton =
           toolButtonOpt->features.testFlag(QStyleOptionToolButton::ToolButtonFeature::MenuButtonPopup);
 
-        const auto& iconSize = _impl->theme.iconSize;
+        const auto& iconSize = toolButtonOpt->iconSize;
         const auto separatorW = _impl->theme.borderWidth;
         const auto spacing = _impl->theme.spacing;
         const auto menuButtonW =
@@ -3677,11 +3725,11 @@ QSize QlementineStyle::sizeFromContents(
     case CT_ToolButton:
       if (const auto* optToolButton = qstyleoption_cast<const QStyleOptionToolButton*>(opt)) {
         const auto spacing = _impl->theme.spacing;
-        const auto& iconSize = _impl->theme.iconSize;
+        const auto& iconSize = optToolButton->iconSize;
 
         // Special cases.
         if (widget->inherits("QLineEditIconButton")) {
-          return iconSize;
+          return _impl->theme.iconSize;
         } else if (widget->inherits("QMenuBarExtension")) {
           const auto extent = pixelMetric(PM_ToolBarExtensionExtent);
           return QSize{ extent, extent };
@@ -3698,7 +3746,8 @@ QSize QlementineStyle::sizeFromContents(
 
         const auto separatorW = menuIsOnSeparateButton ? _impl->theme.borderWidth : 0;
         const auto menuIndicatorW = hasMenu ? separatorW + iconSize.width() + spacing / 2 : 0;
-        const auto h = _impl->theme.controlHeightLarge;
+        const auto h = iconSize.height() < _impl->theme.controlHeightLarge ? _impl->theme.controlHeightLarge
+                                                                           : iconSize.height() + _impl->theme.spacing;
 
         switch (buttonStyle) {
           case Qt::ToolButtonStyle::ToolButtonTextOnly: {
@@ -3733,7 +3782,7 @@ QSize QlementineStyle::sizeFromContents(
     case CT_ComboBox:
       if (const auto* optComboBox = qstyleoption_cast<const QStyleOptionComboBox*>(opt)) {
         // Check if the ComboBox is inside a QTableView/QTreeView.
-        const auto* parentWidget = widget->parentWidget();
+        const auto* parentWidget = widget ? widget->parentWidget() : nullptr;
         const auto* parentParentWidget = parentWidget ? parentWidget->parentWidget() : nullptr;
         const auto isTabCellEditor = qobject_cast<const QAbstractItemView*>(parentParentWidget) != nullptr;
 
@@ -3810,7 +3859,8 @@ QSize QlementineStyle::sizeFromContents(
           // Shortcut. NB: Some difficulties to understand what's going on. Qt changes the width so here's a hack.
           const auto hasShortcut = shortcut.length() > 0;
           const auto reservedShortcutW = optMenuItem->reservedShortcutWidth;
-          const auto shortcutW = hasShortcut ? 3 * spacing - reservedShortcutW : 0;
+          const auto shortcutTextWidth = hasShortcut ? fm.boundingRect(shortcut).width() : 0;
+          const auto shortcutW = std::max(reservedShortcutW, shortcutTextWidth);
 
           // Icon.
           const auto iconW =
@@ -4028,8 +4078,19 @@ int QlementineStyle::pixelMetric(PixelMetric m, const QStyleOption* opt, const Q
     case PM_ButtonIconSize:
       return _impl->theme.iconSize.height();
 
+    // LineEdit.
+    case PM_LineEditIconMargin:
+      return _impl->theme.spacing;
+    case PM_LineEditIconSize:
+      return _impl->theme.iconSize.height();
+
     // Frame.
     case PM_DefaultFrameWidth:
+      // Hack for QLineEdit. This is the only way to know if we have to draw a border or not.
+      // See: https://github.com/qt/qtbase/blob/dev/src/widgets/widgets/qlineedit.cpp#L81C65-L81C85
+      if (qobject_cast<const QLineEdit*>(w)) {
+        return 1;
+      }
       // Prevent QWidgets that contain or inherit QFrame to have a border.
       return 0;
 
@@ -4169,6 +4230,8 @@ int QlementineStyle::pixelMetric(PixelMetric m, const QStyleOption* opt, const Q
       return _impl->theme.spacing * 2;
     case PM_ToolBarExtensionExtent:
       return _impl->theme.iconSize.height() + _impl->theme.spacing;
+    case PM_ToolBarIconSize:
+      return _impl->theme.iconSize.height();
 
     // SpinBox.
     case PM_SpinBoxSliderHeight:
@@ -4178,7 +4241,7 @@ int QlementineStyle::pixelMetric(PixelMetric m, const QStyleOption* opt, const Q
     case PM_IconViewIconSize:
       return pixelMetric(PM_LargeIconSize, opt, w);
     case PM_ListViewIconSize:
-      return _impl->theme.iconSize.height();
+      return pixelMetric(PM_SmallIconSize, opt, w);
     case PM_HeaderDefaultSectionSizeHorizontal:
       return static_cast<int>(_impl->theme.controlDefaultWidth * 1.5);
     case PM_HeaderDefaultSectionSizeVertical:
@@ -4632,6 +4695,9 @@ void QlementineStyle::unpolish(QApplication* app) {
 }
 
 void QlementineStyle::polish(QWidget* w) {
+  if (!w)
+    return;
+
   QCommonStyle::polish(w);
 
 // Currently we only support tooltips with rounded corners on MacOS.
@@ -4656,8 +4722,8 @@ void QlementineStyle::polish(QWidget* w) {
   }
 
   // Prevent the following warning:
-  // QWidget::setMinimumSize: (/QTableCornerButton) Negative sizes (0,-1) are not possible
-  if (w->inherits("QTableCornerButton")) {
+  // QWidget::setMinimumSize: (/QAbstractButton) Negative sizes (0,-1) are not possible
+  if (qobject_cast<QAbstractButton*>(w) && w->minimumSize() == QSize(0, -1)) {
     w->setMinimumSize(0, 1);
   }
 
@@ -4731,9 +4797,15 @@ void QlementineStyle::polish(QWidget* w) {
   }
 
   // Ensure widgets are not compressed vertically.
+  // Some widgets like QCheckBox or QLineEdit are compressed when added to
+  // QFormLayout.
   if (shouldNotBeVerticallyCompressed(w)) {
-    if (0 == w->minimumHeight()) {
-      w->setMinimumHeight(w->sizeHint().height());
+    const auto minHeight = w->minimumHeight();
+    if (minHeight == 0 || minHeight == 1) {
+      const auto heightHint = w->sizeHint().height();
+      if (heightHint > 0) {
+        w->setMinimumHeight(w->sizeHint().height());
+      }
     }
   }
 
@@ -4792,6 +4864,8 @@ void QlementineStyle::polish(QWidget* w) {
     lineEdit->installEventFilter(new LineEditMenuEventFilter(lineEdit));
   } else if (auto* spinBox = qobject_cast<QSpinBox*>(w)) {
     spinBox->installEventFilter(new LineEditMenuEventFilter(spinBox));
+  } else if (auto* plainTextEdit = qobject_cast<QPlainTextEdit*>(w)) {
+    plainTextEdit->installEventFilter(new LineEditMenuEventFilter(plainTextEdit));
   }
 }
 
@@ -5954,15 +6028,15 @@ QColor const& QlementineStyle::tableHeaderBgColor(MouseState const mouse, CheckS
 
   switch (mouse) {
     case MouseState::Pressed:
-      return _impl->theme.neutralColorPressed;
-    case MouseState::Hovered:
       return _impl->theme.neutralColorHovered;
+    case MouseState::Hovered:
+      return _impl->theme.neutralColor;
     case MouseState::Disabled:
       return _impl->theme.neutralColor;
     case MouseState::Transparent:
     case MouseState::Normal:
     default:
-      return _impl->theme.neutralColor;
+      return _impl->theme.backgroundColorMain3;
   }
 }
 
@@ -5976,7 +6050,7 @@ QColor const& QlementineStyle::tableHeaderFgColor(MouseState const mouse, CheckS
 }
 
 QColor const& QlementineStyle::tableLineColor() const {
-  return _impl->theme.secondaryAlternativeColor;
+  return _impl->theme.borderColor;
 }
 
 QColor const& QlementineStyle::colorForTextRole(TextRole role, MouseState const mouse) const {