+++ lib/src/widgets/LoadingSpinner.cpp
@@ -31,15 +31,20 @@ void LoadingSpinner::setSpinning(bool spinning) {
   if (spinning != _spinning) {
     _spinning = spinning;
 
-    if (_spinning) {
+    if (_spinning && isVisible()) {
       _timerId = startTimer(128);
     } else {
-      killTimer(_timerId);
+      if (_timerId != -1) {
+        killTimer(_timerId);
+      }
+      _timerId = -1;
       _i = 0;
     }
 
     emit spinningChanged();
-    update();
+    if (isVisible()) {
+      update();
+    }
   }
 }
 
@@ -90,8 +95,28 @@ void LoadingSpinner::paintEvent(QPaintEvent*) {
   }
 }
 
-void LoadingSpinner::timerEvent(QTimerEvent*) {
-  _i = (_i + 1) % 12;
-  update();
+void LoadingSpinner::timerEvent(QTimerEvent* evt) {
+  if (evt->timerId() == _timerId) {
+    _i = (_i + 1) % 12;
+    if (isVisible()) {
+      update();
+    }
+  }
+}
+
+void LoadingSpinner::showEvent(QShowEvent* evt) {
+  QWidget::showEvent(evt);
+  if (_spinning) {
+    _timerId = startTimer(128);
+  }
+}
+
+void LoadingSpinner::hideEvent(QHideEvent* evt) {
+  QWidget::hideEvent(evt);
+  if (_timerId != -1) {
+    killTimer(_timerId);
+  }
+  _timerId = -1;
+  _i = 0;
 }
 } // namespace oclero::qlementine